the task BJS2-43652 - user registration is ready
