--- src/main/java/school/faang/BJS243652/Main.java
@@ -1,4 +1,0 @@
-package school.faang.BJS243652;
-
-public class Main {
-}
