--- src/main/java/school/faang/BJS243652/User.java
@@ -1,48 +1,0 @@
-package school.faang.BJS243652;
-
-public class User {
-
-    private String name;
-    private int age;
-    private String job;
-    private String address;
-
-    public User(String name, int age, String job, String address) {
-        this.name = name;
-        this.age = age;
-        this.job = job;
-        this.address = address;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public String getJob() {
-        return job;
-    }
-
-    public void setJob(String job) {
-        this.job = job;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-}
