+++ src/main/java/school/faang/task_43652/AllowedAddress.java
@@ -1,0 +1,7 @@
+package school.faang.task_43652;
+
+import java.util.Set;
+
+public class AllowedAddress {
+    public static final Set<String> ALLOWED_ADDRESS = Set.of("London", "New York", "Amsterdam");
+}
