+++ src/main/java/school/faang/task_43652/AllowedJob.java
@@ -1,0 +1,7 @@
+package school.faang.task_43652;
+
+import java.util.Set;
+
+public class AllowedJob {
+    public static final Set<String> ALLOWED_JOB = Set.of("Google", "Uber", "Amazon");
+}
