+++ src/main/java/school/faang/task_43652/Main.java
@@ -1,0 +1,17 @@
+package school.faang.task_43652;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+        User myUser = new User();
+
+        myUser.setName("Anatolijs");
+        myUser.setAge(18);
+        myUser.setJob("Uber");
+        myUser.setAddress("London");
+
+        System.out.println(myUser);
+        System.out.println(myUser.getName());
+    }
+}
